agent_idp_service/pyproject.toml
[project]
name = "agent-idp-service"
version = "0.1.0"
description = "Local MVP Agent Identity Plane on top of Ory/Hydra"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "httpx>=0.28.0",
  "pyjwt>=2.9.0",
  "cryptography>=43.0.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
]

[tool.uv]
package = false
